add custom hook to dispense
