+++ app/src/organisms/ODD/QuickTransferFlow/Dispense/hooks/useDispenseSettingsConfig.ts
@@ -1,0 +1,170 @@
+import { useTranslation } from 'react-i18next'
+import { useToaster } from '/app/organisms/ToasterOven'
+
+import {
+  TRASH_BIN_ADAPTER_FIXTURE,
+  WASTE_CHUTE_FIXTURES,
+} from '@opentrons/shared-data'
+
+import type {
+  DispenseSettingOption,
+  QuickTransferSummaryState,
+  SettingItem,
+} from '../../types'
+
+interface UseDispenseSettingsConfigProps {
+  state: QuickTransferSummaryState
+  setSelectedSetting: (setting: DispenseSettingOption | null) => void
+}
+
+export function useDispenseSettingsConfig({
+  state,
+  setSelectedSetting,
+}: UseDispenseSettingsConfigProps): SettingItem[] {
+  const { t, i18n } = useTranslation(['quick_transfer', 'shared'])
+  const { makeSnackbar } = useToaster()
+
+  const getBlowoutValueCopy = (): string | undefined => {
+    if (state.blowOut === 'dest_well') {
+      return t('blow_out_into_destination_well')
+    } else if (state.blowOut === 'source_well') {
+      return t('blow_out_into_source_well')
+    } else if (
+      state.blowOut != null &&
+      state.blowOut.cutoutFixtureId === TRASH_BIN_ADAPTER_FIXTURE
+    ) {
+      return t('blow_out_into_trash_bin')
+    } else if (
+      state.blowOut != null &&
+      WASTE_CHUTE_FIXTURES.includes(state.blowOut.cutoutFixtureId)
+    ) {
+      return t('blow_out_into_waste_chute')
+    }
+  }
+
+  const dispenseSettingsItems = [
+    {
+      option: 'dispense_flow_rate',
+      copy: t('dispense_flow_rate'),
+      value: t('flow_rate_value', { flow_rate: state.dispenseFlowRate }),
+      enabled: true,
+      onClick: () => {
+        setSelectedSetting('dispense_flow_rate')
+      },
+    },
+    {
+      option: 'dispense_tip_position',
+      copy: t('tip_position'),
+      value:
+        state.tipPositionDispense !== undefined
+          ? t('tip_position_value', { position: state.tipPositionDispense })
+          : '',
+      enabled: true,
+      onClick: () => {
+        setSelectedSetting('dispense_tip_position')
+      },
+    },
+    // ToDo replace dummy configs for submerge
+    {
+      option: 'dispense_submerge',
+      copy: t('submerge'),
+      value: 'dummy Submerge',
+      enabled: false,
+      onClick: () => {
+        // (kk: 04/07/2025)ToDo add submerge setting
+        // setSelectedSetting('submerge')
+      },
+    },
+    {
+      option: 'dispense_delay',
+      copy: t('delay'),
+      value:
+        state.delayDispense !== undefined
+          ? t('delay_value', {
+              delay: state.delayDispense.delayDuration,
+              position: state.delayDispense.positionFromBottom,
+            })
+          : '',
+      enabled: true,
+      onClick: () => {
+        setSelectedSetting('dispense_delay')
+      },
+    },
+    {
+      option: 'dispense_mix',
+      copy: t('mix'),
+      value:
+        state.mixOnDispense !== undefined
+          ? t('mix_value', {
+              volume: state.mixOnDispense?.mixVolume,
+              reps: state.mixOnDispense?.repititions,
+            })
+          : '',
+      enabled:
+        state.transferType === 'transfer' ||
+        state.transferType === 'consolidate',
+      onClick: () => {
+        if (
+          state.transferType === 'transfer' ||
+          state.transferType === 'consolidate'
+        ) {
+          setSelectedSetting('dispense_mix')
+        } else {
+          makeSnackbar(t('advanced_setting_disabled') as string)
+        }
+      },
+    },
+    // ToDo replace dummy configs for push out
+    {
+      option: 'dispense_push_out',
+      copy: t('push_out'),
+      value: 'dummy Push Out',
+      enabled: false,
+      onClick: () => {
+        // (kk: 04/07/2025)ToDo add push out
+        // setSelectedSetting('push_out')
+      },
+    },
+    // ToDo replace dummy configs for retract
+    {
+      option: 'dispense_retract',
+      copy: t('retract'),
+      value: 'dummy Retract',
+      enabled: false,
+      onClick: () => {
+        // (kk: 04/07/2025)ToDo add retract
+        // setSelectedSetting('retract')
+      },
+    },
+    {
+      option: 'dispense_blow_out',
+      copy: t('blow_out'),
+      value:
+        state.transferType === 'distribute'
+          ? t('disabled')
+          : i18n.format(getBlowoutValueCopy(), 'capitalize'),
+      enabled: state.transferType !== 'distribute',
+      onClick: () => {
+        if (state.transferType === 'distribute') {
+          makeSnackbar(t('advanced_setting_disabled') as string)
+        } else {
+          setSelectedSetting('dispense_blow_out')
+        }
+      },
+    },
+    {
+      option: 'dispense_air_gap',
+      copy: t('air_gap'),
+      value:
+        state.airGapDispense !== undefined
+          ? t('air_gap_value', { volume: state.airGapDispense })
+          : '',
+      enabled: true,
+      onClick: () => {
+        setSelectedSetting('dispense_air_gap')
+      },
+    },
+  ]
+
+  return dispenseSettingsItems
+}
