+++ app/src/organisms/ODD/QuickTransferFlow/Dispense/index.tsx
@@ -2,17 +2,13 @@ import { useState, useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { DIRECTION_COLUMN, Flex, SPACING } from '@opentrons/components'
-import {
-  TRASH_BIN_ADAPTER_FIXTURE,
-  WASTE_CHUTE_FIXTURES,
-} from '@opentrons/shared-data'
 
 import { MediumButton } from '/app/atoms/buttons'
 import { ANALYTICS_QUICK_TRANSFER_ADVANCED_SETTINGS_TAB } from '/app/redux/analytics'
 import { useTrackEventWithRobotSerial } from '/app/redux-resources/analytics'
-import { useToaster } from '/app/organisms/ToasterOven'
 import { DispenseSettingDetail } from './DispenseSettingDetail'
 import { DispenseSettingsList } from './DispenseSettingsList'
+import { useDispenseSettingsConfig } from './hooks/useDispenseSettingsConfig'
 
 import type { Dispatch } from 'react'
 import type {
@@ -30,13 +26,16 @@ interface DispenseProps {
 
 export function Dispense(props: DispenseProps): JSX.Element | null {
   const { state, dispatch } = props
-  const { t, i18n } = useTranslation(['quick_transfer', 'shared'])
+  const { t } = useTranslation(['quick_transfer', 'shared'])
   const [
     selectedSetting,
     setSelectedSetting,
   ] = useState<DispenseSettingOption | null>(null)
   const { trackEventWithRobotSerial } = useTrackEventWithRobotSerial()
-  const { makeSnackbar } = useToaster()
+  const dispenseSettingsItems = useDispenseSettingsConfig({
+    state,
+    setSelectedSetting,
+  })
 
   useEffect(() => {
     trackEventWithRobotSerial({
@@ -44,145 +43,6 @@ export function Dispense(props: DispenseProps): JSX.Element | null {
       properties: {},
     })
   }, [])
-
-  function getBlowoutValueCopy(): string | undefined {
-    if (state.blowOut === 'dest_well') {
-      return t('blow_out_into_destination_well')
-    } else if (state.blowOut === 'source_well') {
-      return t('blow_out_into_source_well')
-    } else if (
-      state.blowOut != null &&
-      state.blowOut.cutoutFixtureId === TRASH_BIN_ADAPTER_FIXTURE
-    ) {
-      return t('blow_out_into_trash_bin')
-    } else if (
-      state.blowOut != null &&
-      WASTE_CHUTE_FIXTURES.includes(state.blowOut.cutoutFixtureId)
-    ) {
-      return t('blow_out_into_waste_chute')
-    }
-  }
-
-  const dispenseSettingsItems = [
-    {
-      option: 'dispense_flow_rate',
-      copy: t('dispense_flow_rate'),
-      value: t('flow_rate_value', { flow_rate: state.dispenseFlowRate }),
-      enabled: true,
-      onClick: () => {
-        setSelectedSetting('dispense_flow_rate')
-      },
-    },
-    {
-      option: 'dispense_tip_position',
-      copy: t('tip_position'),
-      value:
-        state.tipPositionDispense !== undefined
-          ? t('tip_position_value', { position: state.tipPositionDispense })
-          : '',
-      enabled: true,
-      onClick: () => {
-        setSelectedSetting('dispense_tip_position')
-      },
-    },
-    // ToDo replace dummy configs for submerge
-    {
-      option: 'dispense_submerge',
-      copy: t('submerge'),
-      enabled: false,
-      onClick: () => {
-        // (kk: 04/07/2025)ToDo add submerge setting
-        // setSelectedSetting('submerge')
-      },
-    },
-    {
-      option: 'dispense_delay',
-      copy: t('delay'),
-      value:
-        state.delayDispense !== undefined
-          ? t('delay_value', {
-              delay: state.delayDispense.delayDuration,
-              position: state.delayDispense.positionFromBottom,
-            })
-          : '',
-      enabled: true,
-      onClick: () => {
-        setSelectedSetting('dispense_delay')
-      },
-    },
-    {
-      option: 'dispense_mix',
-      copy: t('mix'),
-      value:
-        state.mixOnDispense !== undefined
-          ? t('mix_value', {
-              volume: state.mixOnDispense?.mixVolume,
-              reps: state.mixOnDispense?.repititions,
-            })
-          : '',
-      enabled:
-        state.transferType === 'transfer' ||
-        state.transferType === 'consolidate',
-      onClick: () => {
-        if (
-          state.transferType === 'transfer' ||
-          state.transferType === 'consolidate'
-        ) {
-          setSelectedSetting('dispense_mix')
-        } else {
-          makeSnackbar(t('advanced_setting_disabled') as string)
-        }
-      },
-    },
-    // ToDo replace dummy configs for push out
-    {
-      option: 'dispense_push_out',
-      copy: t('push_out'),
-      enabled: false,
-      onClick: () => {
-        // (kk: 04/07/2025)ToDo add push out
-        // setSelectedSetting('push_out')
-      },
-    },
-    // ToDo replace dummy configs for retract
-    {
-      option: 'dispense_retract',
-      copy: t('retract'),
-      enabled: false,
-      onClick: () => {
-        // (kk: 04/07/2025)ToDo add retract
-        // setSelectedSetting('retract')
-      },
-    },
-    {
-      option: 'dispense_blow_out',
-      copy: t('blow_out'),
-      value:
-        state.transferType === 'distribute'
-          ? t('disabled')
-          : i18n.format(getBlowoutValueCopy(), 'capitalize'),
-      enabled: state.transferType !== 'distribute',
-      onClick: () => {
-        if (state.transferType === 'distribute') {
-          makeSnackbar(t('advanced_setting_disabled') as string)
-        } else {
-          setSelectedSetting('dispense_blow_out')
-        }
-      },
-    },
-    {
-      option: 'dispense_air_gap',
-      copy: t('air_gap'),
-      value:
-        state.airGapDispense !== undefined
-          ? t('air_gap_value', { volume: state.airGapDispense })
-          : '',
-      enabled: true,
-      onClick: () => {
-        setSelectedSetting('dispense_air_gap')
-      },
-    },
-  ]
 
   return (
     <Flex
